support queue callback mode
