+++ src/app/(backend)/api/agent/route.ts
@@ -3,40 +3,10 @@ import { type NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
 import { getServerDB } from '@/database/core/db-adaptor';
+import { verifyQStashSignature } from '@/libs/qstash';
 import { AiAgentService } from '@/server/services/aiAgent';
 
 const log = debug('api-route:agent:exec');
-
-/**
- * Verify QStash signature using Receiver
- * Returns true if verification is disabled or signature is valid
- */
-const verifyQStashSignature = async (request: NextRequest, rawBody: string): Promise<boolean> => {
-  const currentSigningKey = process.env.QSTASH_CURRENT_SIGNING_KEY;
-  const nextSigningKey = process.env.QSTASH_NEXT_SIGNING_KEY;
-
-  // If no signing keys configured, skip verification
-  if (!currentSigningKey || !nextSigningKey) {
-    log('QStash signature verification disabled (no signing keys configured)');
-    return false;
-  }
-
-  const signature = request.headers.get('Upstash-Signature');
-  if (!signature) {
-    log('Missing Upstash-Signature header');
-    return false;
-  }
-
-  const { Receiver } = await import('@upstash/qstash');
-  const receiver = new Receiver({ currentSigningKey, nextSigningKey });
-
-  try {
-    return await receiver.verify({ body: rawBody, signature });
-  } catch (error) {
-    log('QStash signature verification failed: %O', error);
-    return false;
-  }
-};
 
 /**
  * Verify API key from Authorization header
