+++ src/libs/qstash/index.ts
@@ -1,5 +1,8 @@
-import { Client } from '@upstash/qstash';
+import { Client, Receiver } from '@upstash/qstash';
 import { Client as WorkflowClient } from '@upstash/workflow';
+import debug from 'debug';
+
+const log = debug('lobe-server:qstash');
 
 const headers = {
   ...(process.env.VERCEL_AUTOMATION_BYPASS_SECRET && {
@@ -26,3 +29,32 @@ export const workflowClient = new WorkflowClient({
   headers,
   token: process.env.QSTASH_TOKEN!,
 });
+
+/**
+ * Verify QStash signature using Receiver.
+ * Returns true if signing keys are not configured (verification skipped) or signature is valid.
+ */
+export async function verifyQStashSignature(request: Request, rawBody: string): Promise<boolean> {
+  const currentSigningKey = process.env.QSTASH_CURRENT_SIGNING_KEY;
+  const nextSigningKey = process.env.QSTASH_NEXT_SIGNING_KEY;
+
+  if (!currentSigningKey || !nextSigningKey) {
+    log('QStash signature verification disabled (no signing keys configured)');
+    return false;
+  }
+
+  const signature = request.headers.get('Upstash-Signature');
+  if (!signature) {
+    log('Missing Upstash-Signature header');
+    return false;
+  }
+
+  const receiver = new Receiver({ currentSigningKey, nextSigningKey });
+
+  try {
+    return await receiver.verify({ body: rawBody, signature });
+  } catch (error) {
+    log('QStash signature verification failed: %O', error);
+    return false;
+  }
+}
